CODE/visualization/datamodel/requirements.txt
kagglehub==1.0.0
python-dotenv==1.2.2
pandas==3.0.2